refactor(admin): split analytics route panels
Why: keep the Quantercise analytics page focused on route and Suspense wiring while moving the remote analytics load and panel rendering into named modules.

Verify: pnpm exec prettier --check apps/admin/src/app/quantercise/analytics/page.tsx apps/admin/src/app/quantercise/analytics/analytics-content.tsx apps/admin/src/app/quantercise/analytics/analytics-sections.tsx; git diff --check -- apps/admin/src/app/quantercise/analytics/page.tsx apps/admin/src/app/quantercise/analytics/analytics-content.tsx apps/admin/src/app/quantercise/analytics/analytics-sections.tsx; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; pnpm validate
